Rename ...types.generator.statechart.core to types.generator.statechart (sort order)
