+++ plugins/org.yakindu.sct.types.generator.statechart.c/META-INF/MANIFEST.MF
@@ -11,7 +11,7 @@ Require-Bundle: org.yakindu.sct.types.generator,
  org.eclipse.xtext.xbase.lib,
  org.eclipse.xtend.lib,
  org.eclipse.xtend.lib.macro,
- org.yakindu.sct.types.generator.statechart.core,
+ org.yakindu.sct.types.generator.statechart,
  org.yakindu.base.types,
  org.yakindu.sct.generator.core,
  org.yakindu.sct.types.modification,
