+++ plugins/org.yakindu.sct.types.generator.statechart.c/src/org/yakindu/sct/types/generator/statechart/c/CSCTGeneratorModule.xtend
@@ -10,12 +10,12 @@
 */
 package org.yakindu.sct.types.generator.statechart.c
 
+import org.yakindu.sct.types.generator.c.CTypesGeneratorModule
+import org.yakindu.sct.types.generator.statechart.SCTGeneratorModule
 import org.yakindu.sct.types.generator.statechart.c.artifacts.CGeneratorArtifactConfigurator
 import org.yakindu.sct.types.generator.statechart.c.naming.CEventNaming
 import org.yakindu.sct.types.generator.statechart.c.naming.CNamingService
 import org.yakindu.sct.types.generator.statechart.c.naming.CPropertyAccessNaming
-import org.yakindu.sct.types.generator.statechart.core.SCTGeneratorModule
-import org.yakindu.sct.types.generator.c.CTypesGeneratorModule
 
 class CSCTGeneratorModule extends SCTGeneratorModule {
