+++ plugins/org.yakindu.sct.types.generator.statechart.core/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.yakindu.sct.types.generator.statechart.core</name>
+	<name>org.yakindu.sct.types.generator.statechart</name>
 	<comment></comment>
 	<projects>
 	</projects>
