+++ plugins/org.yakindu.sct.types.generator.statechart.core/META-INF/MANIFEST.MF
@@ -1,7 +1,7 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Statecharts Code-Generator Core
-Bundle-SymbolicName: org.yakindu.sct.types.generator.statechart.core
+Bundle-SymbolicName: org.yakindu.sct.types.generator.statechart
 Bundle-Version: 3.5.3.qualifier
 Bundle-Vendor: statecharts.org
 Automatic-Module-Name: org.yakindu.sct.generator.slang.core
@@ -14,5 +14,5 @@ Require-Bundle: org.yakindu.sct.types.generator;visibility:=reexport,
  org.yakindu.sct.generator.core,
  org.yakindu.sct.model.sgen,
  org.yakindu.sct.model.sequencer
-Export-Package: org.yakindu.sct.types.generator.statechart.core,
- org.yakindu.sct.types.generator.statechart.core.naming
+Export-Package: org.yakindu.sct.types.generator.statechart,
+ org.yakindu.sct.types.generator.statechart.naming
