+++ plugins/org.yakindu.sct.types.generator.statechart.core/src/org/yakindu/sct/types/generator/statechart/SCTGeneratorModule.java
@@ -8,7 +8,7 @@
  * committers of YAKINDU - initial API and implementation
  *
 */
-package org.yakindu.sct.types.generator.statechart.core;
+package org.yakindu.sct.types.generator.statechart;
 
 import java.util.List;
 
@@ -20,8 +20,8 @@ import org.yakindu.sct.types.generator.ITargetPlatform;
 import org.yakindu.sct.types.generator.ITypesGenerator;
 import org.yakindu.sct.types.generator.artifacts.IGeneratorArtifactConfigurator;
 import org.yakindu.sct.types.generator.module.TypesGeneratorModule;
-import org.yakindu.sct.types.generator.statechart.core.naming.IEventNaming;
-import org.yakindu.sct.types.generator.statechart.core.naming.IPropertyAccessNaming;
+import org.yakindu.sct.types.generator.statechart.naming.IEventNaming;
+import org.yakindu.sct.types.generator.statechart.naming.IPropertyAccessNaming;
 import org.yakindu.sct.types.modification.IModification;
 import org.yakindu.sct.types.modification.ModificationExecutor;
