--- plugins/org.yakindu.sct.types.generator.statechart.core/xtend-gen/org/yakindu/sct/types/generator/statechart/core/.gitignore
@@ -1,2 +1,0 @@
-/.PackageGeneratorEntryExecutor.java._trace
-/PackageGeneratorEntryExecutor.java
