--- plugins/org.yakindu.sct.types.generator.statechart.core/xtend-gen/org/yakindu/sct/types/generator/statechart/core/modification/library/.gitignore
@@ -1,2 +1,0 @@
-/.PropertyAccessModification.java._trace
-/.PropertyReferenceModification.java._trace
